+++ NEWS.md
@@ -211,6 +211,15 @@ cvc5 1.4.0 prerelease
   * LibPoly is now linked statically on Windows, and build triplet detection is
     fixed for Windows on ARM.
 
+- Fixes a bug in the coverings solver, where zero coefficients were ignored from the
+loop that collects relevant coefficients for the projections, whereas they should break
+the loop. (#12926)
+
+- Fixes a bug in the coverings solver, where nullified polynomials were being ignored,
+while they invalidate the guarantees expected by the algorithm. We now detect nullified
+polynomials and, instead of generalizing the sample point to a cell, only exclude the
+sample point itself. If this happens too often, we abort with unknown. (#12943)
+
 cvc5 1.3.4
 ==========